+++ packages/dash-table/src/dash-table/utils/TableClipboardHelper.ts
@@ -42,6 +42,11 @@ export default class TableClipboardHelper {
         Clipboard.set(e, value);
     }
 
+    public static clearClipboard() {
+        TableClipboardHelper.lastLocalCopy = [];
+        TableClipboardHelper.localCopyWithoutHeaders = [];
+    }
+
     public static fromClipboard(
         ev: ClipboardEvent,
         activeCell: ICellCoordinates,